Better files organization
